添加素材文件表

+++ src/YiAim.Cms.Application.Contracts/Blogs/Blog/BaseBlogDto.cs
@@ -4,7 +4,7 @@ using System.Text;
 using Volo.Abp.Application.Dtos;
 
 namespace YiAim.Cms.Blogs;
-public class BaseBlogDto : EntityDto<int>
+public class BaseBlogDto : EntityDto<long>
 {
     public string Title { get; set; }
     public int Taxis { get; set; }
